requirements.txt
asgiref
Django
pytz
sqlparse
psycopg2
djangorestframework
gunicorn
django-cors-headers
django-redis